% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{gre-package}
\alias{gre}
\alias{gre-package}
\title{gre: Generalized Random Effects}
\description{
Computes and returns generalized random effect residuals
}
\author{
\strong{Maintainer}: Benjamin K. Dahl \email{benjamin.dahl@duke.edu} (\href{https://orcid.org/0009-0008-2502-2210}{ORCID})

Other contributors:
\itemize{
  \item Alex Crichton \email{alex@alexcrichton.com} (Rust crate: proc-macro2) [contributor]
  \item David B. Dahl \email{dahl@stat.byu.edu} (Rust crates: roxido, roxido_macro) [contributor]
  \item David Tolnay \email{dtolnay@gmail.com} (Rust crates: proc-macro2, quote, syn, unicode-ident) [contributor]
}

}
\keyword{internal}
