% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_gre.R
\name{initialize_gre}
\alias{initialize_gre}
\title{Initialize a gre External Pointer}
\usage{
initialize_gre(n, m, prior_lambda_shape, prior_lambda_rate)
}
\arguments{
\item{n}{An integer, the number of observations.}

\item{m}{An integer, the number of eigenvalues to consider.}

\item{prior_lambda_shape}{A scalar, by default 2.}

\item{prior_lambda_rate}{A scalar, by default 2.}
}
\value{
An external pointer.
}
\description{
This function initializes an object implementing the method of
(cite the paper here, I guess). The state of the object is stored
internally.
}
\examples{
initialize_gre(1000, 100, 2, 2)
}
