% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run The gre Sampler}
\usage{
run(ptr)
}
\arguments{
\item{ptr}{An external pointer generated by the initialize_gre() function.}
}
\value{
Nothing.
}
\description{
This function samples the correlation adjustment described in (cite).
State is stored internally; nothing is returned.
}
\examples{
run(greObj)
}
