% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_c_eta.R
\name{sample_c_eta}
\alias{sample_c_eta}
\title{Extract Correlation Adjustment}
\usage{
sample_c_eta(ptr)
}
\arguments{
\item{ptr}{An external pointer created by the initialize_gre() function.}
}
\value{
A scalar.
}
\description{
This function extracts the correlation adjustment C times eta in
(cite here too, I guess). When implementing a sampler according to
(cite here too), this vector should be subtracted from the response
before the mean function contribution Gibbs step is entered.
}
\examples{
get_c_eta(greObj)
}
