% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strongarm_update_everything.R
\name{strongarm_update_everything}
\alias{strongarm_update_everything}
\title{Extract Correlation Adjustment}
\usage{
strongarm_update_everything(ptr, kappa, s, lambda, eta, c_eta)
}
\arguments{
\item{ptr}{An external pointer created by the initialize_gre() function.}
}
\value{
A scalar.
}
\description{
This function extracts the correlation adjustment C times eta in
(cite here too, I guess). When implementing a sampler according to
(cite here too), this vector should be subtracted from the response
before the mean function contribution Gibbs step is entered.
}
\examples{
get_c_eta(greObj)
}
