% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_epsilon.R
\name{update_epsilon}
\alias{update_epsilon}
\title{Update (Correlated) Residual}
\usage{
update_epsilon(ptr, new_epsilon)
}
\arguments{
\item{ptr}{An external pointer created by the initialize_gre() function.}

\item{new_epsilon}{A numeric vector of (correlated) residuals.}
}
\value{
Nothing.
}
\description{
This function internally updates a gre external pointer's residual and
adjusts the kappa and s parameters as described in (cite). Nothing is
returned, but the state is stored internally. The length of new_epsilon
must agree with the number of observations the sampler was initialized
to have.
}
\examples{
update_epsilon(greObj, rnorm(100)) # If the sampler has 100 observations
}
